feat(go): improve GetObjects performance and semantics (#2254)
Fixes #2171 

Improves the channel handling and query building for metadata conversion
to Arrow for better performance.

For all cases except when retrieving Column metadata we'll now utilize
`SHOW` queries and build the patterns into those queries. This allows
those `GetObjects` calls with appropriate depths to be called without
having to specify a current database or schema.