feat(go): add new function to allow loading by manifest (#2918)
Following the discussions on the mailing list and on
https://docs.google.com/document/d/1ANeoDI7MCi9xFt_q8z726xKweWS2vgqO8Ut6BSTt5zM/edit?tab=t.0#heading=h.rd6unnsl7uy
this PR implements the proposed behavior by adding a new function to the
driver manager rather than changing the existing AdbcLoadDriver (for
now).

The new function, `AdbcFindLoadDriver` adds a parameter for LoadOptions
along with behavior to search directories at multiple levels
(environment, user, system) to search for either a manifest or a driver
library. This should improve the developer experience and make it easier
for users and applications to leverage the driver manager.

---------

Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: Sutou Kouhei <kou@clear-code.com>
Co-authored-by: Dewey Dunnington <dewey@dunnington.ca>
Co-authored-by: David Li <li.davidm96@gmail.com>